fix: decouple from app version, fix nodetool-core API drift - Update imports: JobUpdate moved from nodetool.types.job (removed   in newer nodetool-core) to nodetool.workflows.types. Without this,   importing nodetool-huggingface against current nodetool-core fails   with ModuleNotFoundError, which silently breaks the Electron stdio   bridge. - Relax pyproject pin: nodetool-core==0.6.3-rc.47 ->   nodetool-core>=0.7.0rc8. The hard equality previously forced uv to   downgrade nodetool-core during install, undoing whatever the app   installer requested. Lower bound only — bump it when relying on a   new nodetool-core API. - Bump version to 0.7.0-rc.8 for the next publish so the registry has   a wheel compatible with the new bridge protocol.
