+++ pyproject.toml
@@ -4,14 +4,18 @@ build-backend = "hatchling.build"
 
 [project]
 name = "nodetool-huggingface"
-version = "0.6.3-rc.47"
+version = "0.7.0-rc.8"
 description = "HuggingFace nodes for Nodetool"
 readme = "README.md"
 authors = [{ name = "Matthias Georgi", email = "matti.georgi@gmail.com" }]
 requires-python = ">=3.11"
 
 dependencies = [
-    "nodetool-core==0.6.3-rc.47",
+    # Lower bound only — let uv resolve to whatever published nodetool-core
+    # works with the rest of the user's environment. Bump this lower bound
+    # only when relying on a new nodetool-core API. See
+    # https://github.com/nodetool-ai/nodetool/blob/main/packages/runtime/src/bridge-protocol.ts
+    "nodetool-core>=0.7.0rc8",
     "torch==2.9.0",
     "nunchaku ; platform_system != 'Darwin'",
     "torchsde",
